refactored tests; added test cases
