+++ tests/Console/ModelsCommand/LaravelCustomCasts/Casts/CustomCasterWithNullablePrimitiveReturn.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts;
+
+use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
+
+class CustomCasterWithNullablePrimitiveReturn implements CastsAttributes
+{
+    /**
+     * @inheritDoc
+     * @return string
+     */
+    public function get($model, string $key, $value, array $attributes): ?array
+    {
+        return new CastedProperty();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function set($model, string $key, $value, array $attributes)
+    {
+        // TODO: Implement set() method.
+    }
+}
